上次commit内容有误补充：1. 完成jwt鉴权及用户解析；2. Entity类重做及配置关系；3. 持久层类字段从驼峰映射到下划线；4. 社交模块完成了：查看好友列表、发送好友申请、响应好友申请、查看发送的申请、查看接收的申请；5.想不起来做了多少了…… \n 本次commit：用户模块新增两个接口：找自己、找朋友

+++ WHUChat/WHUChat.Server/Repositories/IUserRepository.cs
@@ -6,6 +6,6 @@ namespace WHUChat.Server.Repositories
     {
         void Add(User user);
         User GetByUsername(string username);
-        User GetById(int id);
+        User GetById(long id);
     }
 }
