+++ WHUChat/WHUChat.Server/Services/IUserService.cs
@@ -8,7 +8,7 @@ namespace WHUChat.Server.Services
     {
         void Register(RegisterRequestDto request);
         string Login(LoginRequestDto request);
-        User GetById(int id);
+        User GetById(long id);
 
     }
 }
